+++ ICD.Connect.Lighting.Lutron/Properties/AssemblyInfo.cs
@@ -3,5 +3,5 @@ using System.Reflection;
 [assembly: AssemblyTitle("ICD.Connect.Lighting.Lutron")]
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Lighting.Lutron")]
-[assembly: AssemblyCopyright("Copyright © ICD Systems 2017")]
-[assembly: AssemblyVersion("3.0.1.0")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
+[assembly: AssemblyVersion("3.1.2.0")]