+++ ICD.Connect.Lighting.Server/Properties/AssemblyInfo.cs
@@ -3,6 +3,6 @@ using System.Reflection;
 [assembly: AssemblyTitle("ICD.Connect.Lighting.Server")]
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Lighting.Server")]
-[assembly: AssemblyCopyright("Copyright © ICD Systems 2017")]
-[assembly: AssemblyVersion("3.0.1.0")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
+[assembly: AssemblyVersion("3.1.2.0")]